% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/structures.R
\name{Andamento}
\alias{Andamento}
\title{Andamento}
\usage{
Andamento(
  IdAndamento = NULL,
  IdTarefa = NULL,
  IdTarefaModulo = NULL,
  Descricao = NULL,
  DataHora = NULL,
  Unidade = NULL,
  Usuario = NULL,
  Atributos = list()
)
}
\arguments{
\item{IdAndamento}{Character. Internal identifier of the andamento.}

\item{IdTarefa}{Character. Identifier of the associated task.}

\item{IdTarefaModulo}{Character. Identifier of the module task.}

\item{Descricao}{Character. Text describing the andamento.}

\item{DataHora}{Character. Date/time when the andamento was generated.}

\item{Unidade}{A \code{\link{Unidade}} object describing the unit that created the andamento.}

\item{Usuario}{A \code{\link{Usuario}} object describing the user who created the andamento.}

\item{Atributos}{A list of \code{\link{AtributoAndamento}} objects (can be empty).}
}
\value{
An S3 object of class "Andamento".
}
\description{
Represents the "Andamento" structure in SEI, which tracks process events.
}
