% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/structures.R
\name{AndamentoMarcador}
\alias{AndamentoMarcador}
\title{AndamentoMarcador}
\usage{
AndamentoMarcador(
  IdAndamentoMarcador = NULL,
  Texto = NULL,
  DataHora = NULL,
  Usuario = NULL,
  Marcador = NULL
)
}
\arguments{
\item{IdAndamentoMarcador}{Character. Internal ID of the marker event.}

\item{Texto}{Character. Text associated with the andamento marker.}

\item{DataHora}{Character. Date/time the marker was generated.}

\item{Usuario}{A \code{\link{Usuario}} object describing the user who generated the marker.}

\item{Marcador}{A \code{\link{Marcador}} object used in the marker event (or \code{NULL} if removed).}
}
\value{
An S3 object of class "AndamentoMarcador".
}
\description{
Represents the "AndamentoMarcador" structure in SEI (marker events).
}
