% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/escrita.R
\name{anexar_processo}
\alias{anexar_processo}
\title{anexar_processo}
\usage{
anexar_processo(
  protocolo_principal,
  protocolo_anexado,
  config = sei_config(),
  verbose = FALSE
)
}
\arguments{
\item{protocolo_principal, protocolo_anexado}{Números dos processos.}

\item{config}{Um objeto \code{\link[=sei_config]{sei_config()}}.}

\item{verbose}{Logical.}
}
\value{
\code{TRUE} em caso de sucesso.
}
\description{
Anexa um processo a outro (\code{anexarProcesso}).
}
