% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/structures.R
\name{ArquivoExtensao}
\alias{ArquivoExtensao}
\title{ArquivoExtensao}
\usage{
ArquivoExtensao(IdArquivoExtensao, Extensao, Descricao)
}
\arguments{
\item{IdArquivoExtensao}{Character. Internal SEI ID for the extension.}

\item{Extensao}{Character. The extension text (e.g., "pdf", "doc").}

\item{Descricao}{Character. Description of the extension.}
}
\value{
An S3 object of class "ArquivoExtensao".
}
\description{
Represents a file extension allowed in SEI (e.g., pdf, doc).
}
