% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/structures.R
\name{Assinatura}
\alias{Assinatura}
\title{Assinatura}
\usage{
Assinatura(
  Nome = NULL,
  CargoFuncao = NULL,
  DataHora = NULL,
  IdUsuario = NULL,
  IdOrigem = NULL,
  IdOrgao = NULL,
  Sigla = NULL
)
}
\arguments{
\item{Nome}{Character. Signatory name.}

\item{CargoFuncao}{Character. Signatory position or function.}

\item{DataHora}{Character. Date/time of signature.}

\item{IdUsuario}{Character. ID of the user.}

\item{IdOrigem}{Character. Origin ID of the user in the SIP.}

\item{IdOrgao}{Character. ID of the user's organization.}

\item{Sigla}{Character. The user's sigla (username).}
}
\value{
An S3 object of class "Assinatura".
}
\description{
Represents the "Assinatura" structure (signature information).
}
