% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/structures.R
\name{Assunto}
\alias{Assunto}
\title{Assunto}
\usage{
Assunto(CodigoEstruturado = NULL, Descricao = NULL)
}
\arguments{
\item{CodigoEstruturado}{Character. Structured code for the subject (e.g. "00.01.01.01").}

\item{Descricao}{Character. Description of the subject.}
}
\value{
An S3 object of class "Assunto".
}
\description{
Represents the "Assunto" structure in SEI (subject classification).
}
