% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/escrita.R
\name{atribuir_processo}
\alias{atribuir_processo}
\title{atribuir_processo}
\usage{
atribuir_processo(
  protocolo_procedimento,
  id_usuario,
  config = sei_config(),
  sin_reabrir = "N",
  verbose = FALSE
)
}
\arguments{
\item{protocolo_procedimento}{Character. Número do processo.}

\item{id_usuario}{Character. Id do usuário no SIP.}

\item{config}{Um objeto \code{\link[=sei_config]{sei_config()}}.}

\item{sin_reabrir}{"S"/"N".}

\item{verbose}{Logical.}
}
\value{
\code{TRUE} em caso de sucesso.
}
\description{
Atribui um processo a um usuário na unidade (\code{atribuirProcesso}).
}
