% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/structures.R
\name{AtributoAndamento}
\alias{AtributoAndamento}
\title{AtributoAndamento}
\usage{
AtributoAndamento(Nome = NULL, Valor = NULL, IdOrigem = NULL)
}
\arguments{
\item{Nome}{Character. Attribute name.}

\item{Valor}{Character. Attribute value.}

\item{IdOrigem}{Character. Auxiliary ID associated with this attribute.}
}
\value{
An S3 object of class "AtributoAndamento".
}
\description{
Represents the "AtributoAndamento" structure (key-value attributes).
}
