% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/structures.R
\name{AtributoOuvidoria}
\alias{AtributoOuvidoria}
\title{AtributoOuvidoria}
\usage{
AtributoOuvidoria(Id = NULL, Nome = NULL, Titulo = NULL, Valor = NULL)
}
\arguments{
\item{Id}{Character. Identifier of the attribute (e.g., "P").}

\item{Nome}{Character. Internal name of the attribute.}

\item{Titulo}{Character. Title for display in the form.}

\item{Valor}{Character. Value associated with the title.}
}
\value{
An S3 object of class "AtributoOuvidoria".
}
\description{
Represents the "AtributoOuvidoria" structure (attributes in an ouvidoria context).
}
