% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/escrita.R
\name{atualizar_contatos}
\alias{atualizar_contatos}
\title{atualizar_contatos}
\usage{
atualizar_contatos(contatos, config = sei_config(), verbose = FALSE)
}
\arguments{
\item{contatos}{Lista de contatos (objetos \code{\link[=Contato]{Contato()}} / listas nomeadas).}

\item{config}{Um objeto \code{\link[=sei_config]{sei_config()}}.}

\item{verbose}{Logical.}
}
\value{
\code{TRUE} em caso de sucesso.
}
\description{
Atualiza (cadastra/altera) um conjunto de contatos (\code{atualizarContatos}).
}
