% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/soap_helpers.R
\name{build_parameters_xml}
\alias{build_parameters_xml}
\title{Build XML for SOAP Parameters (compatibilidade)}
\usage{
build_parameters_xml(params_list)
}
\arguments{
\item{params_list}{Lista nomeada de parâmetros, ex.:
\code{list(SiglaSistema = "X", IdUnidade = "123")}.}
}
\value{
Uma string XML dos parâmetros.
}
\description{
Converte uma lista nomeada em XML dos parâmetros para o corpo SOAP, no formato
do SEI (com \code{xsi:type="xsd:string"} em escalares, estruturas aninhadas e
arrays \code{<item>}). Delega para o renderizador canônico do pacote.
}
\examples{
build_parameters_xml(list(SiglaSistema = "X", IdUnidade = "123"))

}
\seealso{
\code{\link{sei_build_envelope}}
}
