% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/escrita.R
\name{cancelar_disponibilizacao_bloco}
\alias{cancelar_disponibilizacao_bloco}
\title{cancelar_disponibilizacao_bloco}
\usage{
cancelar_disponibilizacao_bloco(
  id_bloco,
  config = sei_config(),
  verbose = FALSE
)
}
\arguments{
\item{id_bloco}{Character. Número do bloco.}

\item{config}{Um objeto \code{\link[=sei_config]{sei_config()}}.}

\item{verbose}{Logical.}
}
\value{
\code{TRUE} em caso de sucesso.
}
\description{
Cancela a disponibilização de um bloco (\code{cancelarDisponibilizacaoBloco}).
}
