% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/escrita.R
\name{cancelar_documento}
\alias{cancelar_documento}
\title{cancelar_documento}
\usage{
cancelar_documento(
  protocolo_documento,
  motivo,
  config = sei_config(),
  verbose = FALSE
)
}
\arguments{
\item{protocolo_documento}{Character. Número do documento.}

\item{motivo}{Character. Motivo do cancelamento.}

\item{config}{Um objeto \code{\link[=sei_config]{sei_config()}}.}

\item{verbose}{Logical.}
}
\value{
\code{TRUE} em caso de sucesso.
}
\description{
Cancela um documento (\code{cancelarDocumento}).
}
