% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/structures.R
\name{Cargo}
\alias{Cargo}
\title{Cargo}
\usage{
Cargo(
  IdCargo = NULL,
  ExpressaoCargo = NULL,
  ExpressaoTratamento = NULL,
  ExpressaoVocativo = NULL
)
}
\arguments{
\item{IdCargo}{Character. Internal ID of the cargo.}

\item{ExpressaoCargo}{Character. Cargo description (e.g., "Governor").}

\item{ExpressaoTratamento}{Character. Treatment (e.g., "A Sua Excelência o Senhor").}

\item{ExpressaoVocativo}{Character. Vocative (e.g., "Senhor Governador").}
}
\value{
An S3 object of class "Cargo".
}
\description{
Represents a "Cargo" (position) in SEI, with expressions for treatment and vocative.
}
