% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/structures.R
\name{Cidade}
\alias{Cidade}
\title{Cidade}
\usage{
Cidade(
  IdCidade,
  IdEstado,
  IdPais,
  Nome,
  CodigoIbge = NULL,
  SinCapital = NULL,
  Latitude = NULL,
  Longitude = NULL
)
}
\arguments{
\item{IdCidade}{Character. Identifier of the city.}

\item{IdEstado}{Character. Identifier of the state.}

\item{IdPais}{Character. Identifier of the country.}

\item{Nome}{Character. Name of the city.}

\item{CodigoIbge}{Character. IBGE code of the city.}

\item{SinCapital}{"S" or "N", indicates if the city is a state capital.}

\item{Latitude}{Character. Latitude of the city.}

\item{Longitude}{Character. Longitude of the city.}
}
\value{
An S3 object of class "Cidade".
}
\description{
Represents the "Cidade" structure in SEI.
}
