% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/structures.R
\name{DefinicaoControlePrazo}
\alias{DefinicaoControlePrazo}
\title{DefinicaoControlePrazo}
\usage{
DefinicaoControlePrazo(
  ProtocoloProcedimento,
  DataPrazo = NULL,
  Dias = NULL,
  SinDiasUteis = NULL
)
}
\arguments{
\item{ProtocoloProcedimento}{Character. Visible process number, e.g. "12.1.000000077-4".}

\item{DataPrazo}{Character. The date for defining the deadline.}

\item{Dias}{Character. Number of days for the deadline.}

\item{SinDiasUteis}{"S" or "N", indicating if the days are business days.}
}
\value{
An S3 object of class "DefinicaoControlePrazo".
}
\description{
Represents the "DefinicaoControlePrazo" structure for deadline control.
}
