% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/structures.R
\name{DefinicaoMarcador}
\alias{DefinicaoMarcador}
\title{DefinicaoMarcador}
\usage{
DefinicaoMarcador(ProtocoloProcedimento, IdMarcador, Texto = NULL)
}
\arguments{
\item{ProtocoloProcedimento}{Character. Visible process number.}

\item{IdMarcador}{Character. The marker's ID available to the unit.}

\item{Texto}{Character. Text associated with the marker usage.}
}
\value{
An S3 object of class "DefinicaoMarcador".
}
\description{
Represents the "DefinicaoMarcador" structure for applying markers to a process.
}
