% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/escrita.R
\name{definir_controle_prazo}
\alias{definir_controle_prazo}
\title{definir_controle_prazo}
\usage{
definir_controle_prazo(definicoes, config = sei_config(), verbose = FALSE)
}
\arguments{
\item{definicoes}{Lista de definições (cada uma com \code{ProtocoloProcedimento} e
\code{DataPrazo} OU \code{Dias}/\code{SinDiasUteis}).}

\item{config}{Um objeto \code{\link[=sei_config]{sei_config()}}.}

\item{verbose}{Logical.}
}
\value{
\code{TRUE} em caso de sucesso.
}
\description{
Define controle de prazo para processos (\code{definirControlePrazo}).
}
