% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/escrita.R
\name{definir_marcador}
\alias{definir_marcador}
\title{definir_marcador}
\usage{
definir_marcador(definicoes, config = sei_config(), verbose = FALSE)
}
\arguments{
\item{definicoes}{Lista de definições (cada uma \code{list(ProtocoloProcedimento=, IdMarcador=, Texto=)}).}

\item{config}{Um objeto \code{\link[=sei_config]{sei_config()}}.}

\item{verbose}{Logical.}
}
\value{
\code{TRUE} em caso de sucesso.
}
\description{
Define marcadores em processos (\code{definirMarcador}).
}
