% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/escrita.R
\name{desanexar_processo}
\alias{desanexar_processo}
\title{desanexar_processo}
\usage{
desanexar_processo(
  protocolo_principal,
  protocolo_anexado,
  motivo,
  config = sei_config(),
  verbose = FALSE
)
}
\arguments{
\item{protocolo_principal, protocolo_anexado}{Números dos processos.}

\item{motivo}{Character. Motivo da desanexação.}

\item{config}{Um objeto \code{\link[=sei_config]{sei_config()}}.}

\item{verbose}{Logical.}
}
\value{
\code{TRUE} em caso de sucesso.
}
\description{
Desanexa um processo (\code{desanexarProcesso}).
}
