% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/escrita.R
\name{desbloquear_processo}
\alias{desbloquear_processo}
\title{desbloquear_processo}
\usage{
desbloquear_processo(
  protocolo_procedimento,
  config = sei_config(),
  verbose = FALSE
)
}
\arguments{
\item{protocolo_procedimento}{Character. Número do processo.}

\item{config}{Um objeto \code{\link[=sei_config]{sei_config()}}.}

\item{verbose}{Logical.}
}
\value{
\code{TRUE} em caso de sucesso.
}
\description{
Desbloqueia um processo (\code{desbloquearProcesso}).
}
