% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/structures.R
\name{Destinatario}
\alias{Destinatario}
\title{Destinatario}
\usage{
Destinatario(Sigla = NULL, Nome = NULL)
}
\arguments{
\item{Sigla}{Character. The sigla of the recipient.}

\item{Nome}{Character. The name of the recipient.}
}
\value{
An S3 object of class "Destinatario".
}
\description{
Represents the "Destinatario" structure in SEI.
}
