% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/escrita.R
\name{disponibilizar_bloco}
\alias{disponibilizar_bloco}
\title{disponibilizar_bloco}
\usage{
disponibilizar_bloco(id_bloco, config = sei_config(), verbose = FALSE)
}
\arguments{
\item{id_bloco}{Character. Número do bloco.}

\item{config}{Um objeto \code{\link[=sei_config]{sei_config()}}.}

\item{verbose}{Logical.}
}
\value{
\code{TRUE} em caso de sucesso.
}
\description{
Disponibiliza um bloco (\code{disponibilizarBloco}).
}
