% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/escrita.R
\name{enviar_email}
\alias{enviar_email}
\title{enviar_email}
\usage{
enviar_email(
  protocolo_procedimento,
  de,
  para,
  assunto,
  mensagem,
  cco = NULL,
  id_documentos = NULL,
  config = sei_config(),
  raw = FALSE,
  verbose = FALSE
)
}
\arguments{
\item{protocolo_procedimento}{Character. Número do processo.}

\item{de}{Character. Endereço do remetente.}

\item{para}{Vetor de destinatários (serão unidos por ";").}

\item{assunto, mensagem}{Character.}

\item{cco}{Vetor de destinatários em cópia oculta (opcional).}

\item{id_documentos}{Vetor de ids de documentos a anexar (opcional).}

\item{config}{Um objeto \code{\link[=sei_config]{sei_config()}}.}

\item{raw, verbose}{Logical.}
}
\value{
Um \code{tibble} (\link{parse_retorno_envio_email}).
}
\description{
Envia um e-mail vinculado a um processo (\code{enviarEmail}).
}
