% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/escrita.R
\name{enviar_processo}
\alias{enviar_processo}
\title{enviar_processo}
\usage{
enviar_processo(
  protocolo_procedimento,
  unidades_destino,
  config = sei_config(),
  sin_manter_aberto_unidade = "N",
  sin_remover_anotacao = "N",
  sin_enviar_email_notificacao = "N",
  data_retorno_programado = NULL,
  dias_retorno_programado = NULL,
  sin_dias_uteis_retorno_programado = "N",
  sin_reabrir = "N",
  verbose = FALSE
)
}
\arguments{
\item{protocolo_procedimento}{Character. Número do processo.}

\item{unidades_destino}{Vetor de ids de unidades de destino.}

\item{config}{Um objeto \code{\link[=sei_config]{sei_config()}}.}

\item{sin_manter_aberto_unidade, sin_remover_anotacao, sin_enviar_email_notificacao}{"S"/"N".}

\item{data_retorno_programado, dias_retorno_programado, sin_dias_uteis_retorno_programado}{Retorno programado.}

\item{sin_reabrir}{"S"/"N".}

\item{verbose}{Logical.}
}
\value{
\code{TRUE} em caso de sucesso.
}
\description{
Envia (tramita) um processo para outras unidades (\code{enviarProcesso}).
}
