% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/structures.R
\name{Estado}
\alias{Estado}
\title{Estado}
\usage{
Estado(IdEstado, IdPais, Sigla, Nome, CodigoIbge = NULL)
}
\arguments{
\item{IdEstado}{Character. ID of the state.}

\item{IdPais}{Character. ID of the country.}

\item{Sigla}{Character. State sigla.}

\item{Nome}{Character. Name of the state.}

\item{CodigoIbge}{Character. IBGE code of the state.}
}
\value{
An S3 object of class "Estado".
}
\description{
Represents the "Estado" structure in SEI (state info).
}
