% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/escrita.R
\name{excluir_bloco}
\alias{excluir_bloco}
\title{excluir_bloco}
\usage{
excluir_bloco(id_bloco, config = sei_config(), verbose = FALSE)
}
\arguments{
\item{id_bloco}{Character. Número do bloco.}

\item{config}{Um objeto \code{\link[=sei_config]{sei_config()}}.}

\item{verbose}{Logical.}
}
\value{
\code{TRUE} em caso de sucesso.
}
\description{
Exclui um bloco (\code{excluirBloco}).
}
