% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/escrita.R
\name{excluir_documento}
\alias{excluir_documento}
\title{excluir_documento}
\usage{
excluir_documento(protocolo_documento, config = sei_config(), verbose = FALSE)
}
\arguments{
\item{protocolo_documento}{Character. Número do documento.}

\item{config}{Um objeto \code{\link[=sei_config]{sei_config()}}.}

\item{verbose}{Logical.}
}
\value{
\code{TRUE} em caso de sucesso.
}
\description{
Exclui um documento (\code{excluirDocumento}).
}
