% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/escrita.R
\name{excluir_processo}
\alias{excluir_processo}
\title{excluir_processo}
\usage{
excluir_processo(
  protocolo_procedimento,
  config = sei_config(),
  verbose = FALSE
)
}
\arguments{
\item{protocolo_procedimento}{Character. Número do processo.}

\item{config}{Um objeto \code{\link[=sei_config]{sei_config()}}.}

\item{verbose}{Logical.}
}
\value{
\code{TRUE} em caso de sucesso.
}
\description{
Exclui um processo (\code{excluirProcesso}). Irreversível.
}
