% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/structures.R
\name{Feriado}
\alias{Feriado}
\title{Feriado}
\usage{
Feriado(Data, Descricao)
}
\arguments{
\item{Data}{Character. Date of the holiday.}

\item{Descricao}{Character. Description of the holiday.}
}
\value{
An S3 object of class "Feriado".
}
\description{
Represents the "Feriado" structure (holiday).
}
