% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/escrita.R
\name{gerar_bloco}
\alias{gerar_bloco}
\title{gerar_bloco}
\usage{
gerar_bloco(
  tipo,
  descricao,
  config = sei_config(),
  unidades_disponibilizacao = NULL,
  documentos = NULL,
  sin_disponibilizar = "N",
  verbose = FALSE
)
}
\arguments{
\item{tipo}{Character. "A" (assinatura), "R" (reunião) ou "I" (interno).}

\item{descricao}{Character. Descrição do bloco.}

\item{config}{Um objeto \code{\link[=sei_config]{sei_config()}}.}

\item{unidades_disponibilizacao}{Vetor de ids de unidades (ou \code{NULL}).}

\item{documentos}{Vetor de protocolos de documentos (ou \code{NULL}).}

\item{sin_disponibilizar}{"S"/"N".}

\item{verbose}{Logical.}
}
\value{
Character com o \code{IdBloco} gerado.
}
\description{
Gera um bloco (\code{gerarBloco}). Retorna o número do bloco.
}
