% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/escrita.R
\name{gerar_procedimento}
\alias{gerar_procedimento}
\title{gerar_procedimento}
\usage{
gerar_procedimento(
  procedimento,
  config = sei_config(),
  documentos = list(),
  procedimentos_relacionados = NULL,
  unidades_envio = NULL,
  sin_manter_aberto_unidade = "S",
  sin_enviar_email_notificacao = "N",
  data_retorno_programado = NULL,
  dias_retorno_programado = NULL,
  sin_dias_uteis_retorno_programado = "N",
  id_marcador = NULL,
  texto_marcador = NULL,
  data_controle_prazo = NULL,
  dias_controle_prazo = NULL,
  sin_dias_uteis_controle_prazo = "N",
  raw = FALSE,
  verbose = FALSE
)
}
\arguments{
\item{procedimento}{Lista nomeada / objeto \code{\link[=Procedimento]{Procedimento()}} com os dados do processo.}

\item{config}{Um objeto \code{\link[=sei_config]{sei_config()}}.}

\item{documentos}{Lista de documentos a gerar junto (objetos \code{\link[=Documento]{Documento()}}); ou \code{list()}.}

\item{procedimentos_relacionados, unidades_envio}{Vetores de ids (opcionais).}

\item{sin_manter_aberto_unidade, sin_enviar_email_notificacao}{"S"/"N".}

\item{data_retorno_programado, dias_retorno_programado, sin_dias_uteis_retorno_programado}{Retorno programado.}

\item{id_marcador, texto_marcador}{Marcador opcional.}

\item{data_controle_prazo, dias_controle_prazo, sin_dias_uteis_controle_prazo}{Controle de prazo opcional.}

\item{raw, verbose}{Logical.}
}
\value{
Um \code{tibble} (\link{parse_retorno_geracao_procedimento}) ou \code{xml_document} se \code{raw=TRUE}.
}
\description{
Gera um novo processo (\code{gerarProcedimento}). Operação de escrita.
}
