% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_response.R
\name{get_text}
\alias{get_text}
\title{get_text: Extract Text from an XML Node}
\usage{
get_text(parent, child_name)
}
\arguments{
\item{parent}{An \code{xml2} node.}

\item{child_name}{Character. The local-name of the child element to extract.}
}
\value{
A character string (possibly empty).
}
\description{
A helper function to find the first child node of \code{parent} with a given
local-name, ignoring namespaces. It returns the text content unless
\code{xsi:nil="true"}, in which case it returns an empty string.
}
\examples{
\dontrun{
  doc <- xml2::read_xml("<root><Foo xsi:nil='true'/></root>")
  get_text(doc, "Foo")  # => ""
}

}
