% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/structures.R
\name{HipoteseLegal}
\alias{HipoteseLegal}
\title{HipoteseLegal}
\usage{
HipoteseLegal(IdHipoteseLegal, Nome, BaseLegal, NivelAcesso)
}
\arguments{
\item{IdHipoteseLegal}{Character. ID of the legal hypothesis.}

\item{Nome}{Character. Name of the legal hypothesis.}

\item{BaseLegal}{Character. Description of the base legal.}

\item{NivelAcesso}{Character. "1"=restricted, "2"=secret, etc.}
}
\value{
An S3 object of class "HipoteseLegal".
}
\description{
Represents the "HipoteseLegal" structure (legal hypothesis for restricted/sigilo).
}
