% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/escrita.R
\name{incluir_documento}
\alias{incluir_documento}
\title{incluir_documento}
\usage{
incluir_documento(
  documento,
  config = sei_config(),
  raw = FALSE,
  verbose = FALSE
)
}
\arguments{
\item{documento}{Lista nomeada / objeto \code{\link[=Documento]{Documento()}}.}

\item{config}{Um objeto \code{\link[=sei_config]{sei_config()}}.}

\item{raw, verbose}{Logical.}
}
\value{
Um \code{tibble} (\link{parse_retorno_inclusao_documento}) ou \code{xml_document}.
}
\description{
Inclui um documento em um processo (\code{incluirDocumento}). Escrita.
}
