% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/escrita.R
\name{incluir_documento_bloco}
\alias{incluir_documento_bloco}
\title{incluir_documento_bloco}
\usage{
incluir_documento_bloco(
  id_bloco,
  protocolo_documento,
  anotacao = NULL,
  config = sei_config(),
  verbose = FALSE
)
}
\arguments{
\item{id_bloco}{Character. Número do bloco.}

\item{protocolo_documento}{Character. Número do documento.}

\item{anotacao}{Character opcional.}

\item{config}{Um objeto \code{\link[=sei_config]{sei_config()}}.}

\item{verbose}{Logical.}
}
\value{
\code{TRUE} em caso de sucesso.
}
\description{
Inclui um documento em um bloco (\code{incluirDocumentoBloco}).
}
