% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/escrita.R
\name{incluir_processo_bloco}
\alias{incluir_processo_bloco}
\title{incluir_processo_bloco}
\usage{
incluir_processo_bloco(
  id_bloco,
  protocolo_procedimento,
  anotacao = NULL,
  config = sei_config(),
  verbose = FALSE
)
}
\arguments{
\item{id_bloco}{Character. Número do bloco.}

\item{protocolo_procedimento}{Character. Número do processo.}

\item{anotacao}{Character opcional.}

\item{config}{Um objeto \code{\link[=sei_config]{sei_config()}}.}

\item{verbose}{Logical.}
}
\value{
\code{TRUE} em caso de sucesso.
}
\description{
Inclui um processo em um bloco (\code{incluirProcessoBloco}).
}
