% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/structures.R
\name{Interessado}
\alias{Interessado}
\title{Interessado}
\usage{
Interessado(Sigla = NULL, Nome = NULL)
}
\arguments{
\item{Sigla}{Character. The sigla of the interested party.}

\item{Nome}{Character. The name of the interested party.}
}
\value{
An S3 object of class "Interessado".
}
\description{
Represents the "Interessado" structure in SEI (interested party in a process or document).
}
