% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/listas.R
\name{listar_cargos}
\alias{listar_cargos}
\title{listar_cargos}
\usage{
listar_cargos(
  config = sei_config(),
  id_cargo = NULL,
  raw = FALSE,
  verbose = FALSE
)
}
\arguments{
\item{config}{Um objeto \code{\link[=sei_config]{sei_config()}}.}

\item{id_cargo}{Filtro opcional.}

\item{raw, verbose}{Logical.}
}
\value{
Um \code{tibble} de cargos.
}
\description{
Lista os cargos.
}
