% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/listas.R
\name{listar_cidades}
\alias{listar_cidades}
\title{listar_cidades}
\usage{
listar_cidades(
  config = sei_config(),
  id_pais = NULL,
  id_estado = NULL,
  raw = FALSE,
  verbose = FALSE
)
}
\arguments{
\item{config}{Um objeto \code{\link[=sei_config]{sei_config()}}.}

\item{id_pais, id_estado}{Filtros opcionais.}

\item{raw, verbose}{Logical.}
}
\value{
Um \code{tibble} de cidades.
}
\description{
Lista as cidades.
}
