% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/listas.R
\name{listar_contatos}
\alias{listar_contatos}
\title{listar_contatos}
\usage{
listar_contatos(
  config = sei_config(),
  id_tipo_contato = NULL,
  pagina_registros = NULL,
  pagina_atual = NULL,
  sigla = NULL,
  nome = NULL,
  cpf = NULL,
  cnpj = NULL,
  matricula = NULL,
  raw = FALSE,
  verbose = FALSE
)
}
\arguments{
\item{config}{Um objeto \code{\link[=sei_config]{sei_config()}}.}

\item{id_tipo_contato, sigla, nome, cpf, cnpj, matricula}{Filtros opcionais.}

\item{pagina_registros, pagina_atual}{Paginação (1-1000; padrão 1).}

\item{raw, verbose}{Logical.}
}
\value{
Um \code{tibble} de contatos.
}
\description{
Lista contatos (paginado).
}
