% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/listas.R
\name{listar_estados}
\alias{listar_estados}
\title{listar_estados}
\usage{
listar_estados(
  config = sei_config(),
  id_pais = NULL,
  raw = FALSE,
  verbose = FALSE
)
}
\arguments{
\item{config}{Um objeto \code{\link[=sei_config]{sei_config()}}.}

\item{id_pais}{Filtro opcional.}

\item{raw, verbose}{Logical.}
}
\value{
Um \code{tibble} de estados.
}
\description{
Lista os estados (UF).
}
