% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/listas.R
\name{listar_extensoes_permitidas}
\alias{listar_extensoes_permitidas}
\title{listar_extensoes_permitidas}
\usage{
listar_extensoes_permitidas(
  config = sei_config(),
  id_arquivo_extensao = NULL,
  raw = FALSE,
  verbose = FALSE
)
}
\arguments{
\item{config}{Um objeto \code{\link[=sei_config]{sei_config()}}.}

\item{id_arquivo_extensao}{Filtro opcional.}

\item{raw, verbose}{Logical.}
}
\value{
Um \code{tibble} de extensões.
}
\description{
Lista as extensões de arquivo permitidas.
}
