% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/listas.R
\name{listar_feriados}
\alias{listar_feriados}
\title{listar_feriados}
\usage{
listar_feriados(
  config = sei_config(),
  id_orgao = NULL,
  data_inicial = NULL,
  data_final = NULL,
  raw = FALSE,
  verbose = FALSE
)
}
\arguments{
\item{config}{Um objeto \code{\link[=sei_config]{sei_config()}}.}

\item{id_orgao, data_inicial, data_final}{Filtros opcionais.}

\item{raw, verbose}{Logical.}
}
\value{
Um \code{tibble} de feriados.
}
\description{
Lista os feriados.
}
