% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/listas.R
\name{listar_hipoteses_legais}
\alias{listar_hipoteses_legais}
\title{listar_hipoteses_legais}
\usage{
listar_hipoteses_legais(
  config = sei_config(),
  nivel_acesso = NULL,
  raw = FALSE,
  verbose = FALSE
)
}
\arguments{
\item{config}{Um objeto \code{\link[=sei_config]{sei_config()}}.}

\item{nivel_acesso}{Filtro opcional (1 - restrito, 2 - sigiloso).}

\item{raw, verbose}{Logical.}
}
\value{
Um \code{tibble} de hipóteses legais.
}
\description{
Lista as hipóteses legais.
}
