% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/listas.R
\name{listar_marcadores_unidade}
\alias{listar_marcadores_unidade}
\title{listar_marcadores_unidade}
\usage{
listar_marcadores_unidade(config = sei_config(), raw = FALSE, verbose = FALSE)
}
\arguments{
\item{config}{Um objeto \code{\link[=sei_config]{sei_config()}}.}

\item{raw, verbose}{Logical.}
}
\value{
Um \code{tibble} de marcadores.
}
\description{
Lista os marcadores da unidade.
}
