% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/listas.R
\name{listar_paises}
\alias{listar_paises}
\title{listar_paises}
\usage{
listar_paises(config = sei_config(), raw = FALSE, verbose = FALSE)
}
\arguments{
\item{config}{Um objeto \code{\link[=sei_config]{sei_config()}}.}

\item{raw, verbose}{Logical.}
}
\value{
Um \code{tibble} de países.
}
\description{
Lista os países.
}
