% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sip.R
\name{listar_permissao}
\alias{listar_permissao}
\title{listar_permissao}
\usage{
listar_permissao(
  config = sip_config(),
  id_orgao_usuario = NULL,
  id_usuario = NULL,
  id_origem_usuario = NULL,
  id_orgao_unidade = NULL,
  id_unidade = NULL,
  id_origem_unidade = NULL,
  id_perfil = NULL,
  raw = FALSE,
  verbose = FALSE
)
}
\arguments{
\item{config}{Um objeto \code{\link[=sip_config]{sip_config()}}.}

\item{id_orgao_usuario, id_usuario, id_origem_usuario, id_orgao_unidade, id_unidade, id_origem_unidade, id_perfil}{Filtros opcionais.}

\item{raw, verbose}{Logical.}
}
\value{
Um \code{tibble} de permissões (ou \code{xml_document} se \code{raw = TRUE}).
}
\description{
Lista permissões no SIP (\code{listarPermissao}).
}
