% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/listas.R
\name{listar_series}
\alias{listar_series}
\title{listar_series}
\usage{
listar_series(
  config = sei_config(),
  id_unidade = NULL,
  id_tipo_procedimento = NULL,
  raw = FALSE,
  verbose = FALSE
)
}
\arguments{
\item{config}{Um objeto \code{\link[=sei_config]{sei_config()}}.}

\item{id_unidade, id_tipo_procedimento}{Filtros opcionais.}

\item{raw, verbose}{Logical.}
}
\value{
Um \code{tibble} de séries.
}
\description{
Lista os tipos de documento (séries) liberados ao serviço.
}
