% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/listas.R
\name{listar_tipos_conferencia}
\alias{listar_tipos_conferencia}
\title{listar_tipos_conferencia}
\usage{
listar_tipos_conferencia(
  config = sei_config(),
  id_unidade = NULL,
  raw = FALSE,
  verbose = FALSE
)
}
\arguments{
\item{config}{Um objeto \code{\link[=sei_config]{sei_config()}}.}

\item{id_unidade}{Filtro opcional.}

\item{raw, verbose}{Logical.}
}
\value{
Um \code{tibble} de tipos de conferência.
}
\description{
Lista os tipos de conferência.
}
