% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/listas.R
\name{listar_tipos_procedimento}
\alias{listar_tipos_procedimento}
\title{listar_tipos_procedimento}
\usage{
listar_tipos_procedimento(
  config = sei_config(),
  id_unidade = NULL,
  id_serie = NULL,
  raw = FALSE,
  verbose = FALSE
)
}
\arguments{
\item{config}{Um objeto \code{\link[=sei_config]{sei_config()}}.}

\item{id_unidade, id_serie}{Filtros opcionais.}

\item{raw, verbose}{Logical.}
}
\value{
Um \code{tibble} de tipos de processo.
}
\description{
Lista os tipos de processo liberados ao serviço.
}
