% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/listas.R
\name{listar_tipos_procedimento_ouvidoria}
\alias{listar_tipos_procedimento_ouvidoria}
\title{listar_tipos_procedimento_ouvidoria}
\usage{
listar_tipos_procedimento_ouvidoria(
  config = sei_config(),
  raw = FALSE,
  verbose = FALSE
)
}
\arguments{
\item{config}{Um objeto \code{\link[=sei_config]{sei_config()}}.}

\item{raw, verbose}{Logical.}
}
\value{
Um \code{tibble} de tipos de processo.
}
\description{
Lista os tipos de processo sinalizados como de Ouvidoria.
}
