% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/listas.R
\name{listar_usuarios}
\alias{listar_usuarios}
\title{listar_usuarios}
\usage{
listar_usuarios(
  config = sei_config(),
  id_usuario = NULL,
  raw = FALSE,
  verbose = FALSE
)
}
\arguments{
\item{config}{Um objeto \code{\link[=sei_config]{sei_config()}}.}

\item{id_usuario}{Filtro opcional.}

\item{raw, verbose}{Logical.}
}
\value{
Um \code{tibble} de usuários.
}
\description{
Lista os usuários com perfil "Básico" na unidade.
}
