% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/structures.R
\name{Marcador}
\alias{Marcador}
\title{Marcador}
\usage{
Marcador(IdMarcador, Nome, Icone = NULL, SinAtivo = NULL)
}
\arguments{
\item{IdMarcador}{Character. ID of the marker.}

\item{Nome}{Character. Name of the marker.}

\item{Icone}{Character. PNG icon (Base64-encoded).}

\item{SinAtivo}{"S" or "N" indicating if the marker is active.}
}
\value{
An S3 object of class "Marcador".
}
\description{
Represents the "Marcador" structure (markers associated with a process).
}
