% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/structures.R
\name{Observacao}
\alias{Observacao}
\title{Observacao}
\usage{
Observacao(Descricao = NULL, Unidade = NULL)
}
\arguments{
\item{Descricao}{Character. The text of the observation.}

\item{Unidade}{A \code{\link{Unidade}} object that added this observation.}
}
\value{
An S3 object of class "Observacao".
}
\description{
Represents the "Observacao" structure in SEI (unit notes).
}
