% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/structures.R
\name{Pais}
\alias{Pais}
\title{Pais}
\usage{
Pais(IdPais, Nome)
}
\arguments{
\item{IdPais}{Character. The ID of the country.}

\item{Nome}{Character. The name of the country.}
}
\value{
An S3 object of class "Pais".
}
\description{
Represents the "Pais" structure in SEI (country).
}
