% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_escrita.R
\name{parse_andamento_response}
\alias{parse_andamento_response}
\title{parse_andamento_response}
\usage{
parse_andamento_response(doc)
}
\arguments{
\item{doc}{Um \code{xml2::xml_document}.}
}
\value{
Um tibble de 1 linha (ver \link{parse_andamento_item}).
}
\description{
Parseia o retorno de \code{lancarAndamento} (um \code{Andamento}).
}
