% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_listas.R
\name{parse_arquivo_extensao}
\alias{parse_arquivo_extensao}
\title{parse_arquivo_extensao}
\usage{
parse_arquivo_extensao(node)
}
\arguments{
\item{node}{Nó \code{xml2}.}
}
\value{
Um tibble de 1 linha.
}
\description{
Parseia uma estrutura \code{ArquivoExtensao}.
}
