% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_consultas.R
\name{parse_assinatura}
\alias{parse_assinatura}
\title{parse_assinatura}
\usage{
parse_assinatura(node)
}
\arguments{
\item{node}{Nó \code{xml2} da \code{Assinatura} (filho direto ou \verb{<item>}).}
}
\value{
Um tibble de 1 linha com os campos da assinatura.
}
\description{
Parseia uma estrutura \code{Assinatura}.
}
