% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_response.R
\name{parse_assunto}
\alias{parse_assunto}
\title{parse_assunto}
\usage{
parse_assunto(node_ass)
}
\arguments{
\item{node_ass}{An \code{xml2} node for <item> in <Assuntos>.}
}
\value{
A named list with \code{$CodigoEstruturado} and \code{$Descricao}.
}
\description{
Parses one <Assunto> item, extracting \code{CodigoEstruturado} and \code{Descricao}.
}
\examples{
\dontrun{
  parse_assunto(xml_item)
}

}
