% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_consultas.R
\name{parse_campo}
\alias{parse_campo}
\title{parse_campo}
\usage{
parse_campo(node)
}
\arguments{
\item{node}{Nó \code{xml2} do \code{Campo}.}
}
\value{
Um tibble de 1 linha com \code{Nome} e \code{Valor}.
}
\description{
Parseia uma estrutura \code{Campo} (formulário).
}
