% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_listas.R
\name{parse_cargo}
\alias{parse_cargo}
\title{parse_cargo}
\usage{
parse_cargo(node)
}
\arguments{
\item{node}{Nó \code{xml2}.}
}
\value{
Um tibble de 1 linha.
}
\description{
Parseia uma estrutura \code{Cargo}.
}
