% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_listas.R
\name{parse_cidade}
\alias{parse_cidade}
\title{parse_cidade}
\usage{
parse_cidade(node)
}
\arguments{
\item{node}{Nó \code{xml2}.}
}
\value{
Um tibble de 1 linha.
}
\description{
Parseia uma estrutura \code{Cidade}.
}
