% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_consultas.R
\name{parse_consultar_procedimento_individual_response}
\alias{parse_consultar_procedimento_individual_response}
\title{parse_consultar_procedimento_individual_response}
\usage{
parse_consultar_procedimento_individual_response(doc)
}
\arguments{
\item{doc}{Um \code{xml2::xml_document}.}
}
\value{
Um tibble de 1 linha.
}
\description{
Parseia a resposta de \code{consultarProcedimentoIndividual}
(\code{ProcedimentoResumido}). Retorna linha de \code{NA} se nenhum processo for
encontrado (\code{parametros} nulo).
}
