% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_response.R
\name{parse_consultar_procedimento_response}
\alias{parse_consultar_procedimento_response}
\title{parse_consultar_procedimento_response}
\usage{
parse_consultar_procedimento_response(doc)
}
\arguments{
\item{doc}{Um \code{xml2::xml_document} com a resposta de "consultarProcedimento".}
}
\value{
Um tibble de 1 linha. Campos escalares e subestruturas 1:1
(TipoProcedimento e andamentos) viram colunas com prefixo; arrays
(\code{Assuntos}, \code{Interessados}, \code{Observacoes},
\code{UnidadesProcedimentoAberto}, \code{ProcedimentosRelacionados},
\code{ProcedimentosAnexados}) ficam como colunas-lista de tibbles.
}
\description{
Parser principal de \code{consultarProcedimento}: localiza o nó
\code{<parametros>}, extrai os campos escalares, recodifica
\code{NivelAcessoLocal}/\code{NivelAcessoGlobal} e parseia as subestruturas
(\code{TipoProcedimento}, andamentos, e os arrays \code{Assuntos},
\code{Interessados}, etc.).
}
\examples{
\dontrun{
  doc <- consultar_procedimento("0000000000.000001/2020-11", raw = TRUE)
  parse_consultar_procedimento_response(doc)
}

}
