% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_listas.R
\name{parse_contato}
\alias{parse_contato}
\title{parse_contato}
\usage{
parse_contato(node)
}
\arguments{
\item{node}{Nó \code{xml2}.}
}
\value{
Um tibble de 1 linha.
}
\description{
Parseia uma estrutura \code{Contato} (subconjunto dos campos úteis).
}
