% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_listas.R
\name{parse_feriado}
\alias{parse_feriado}
\title{parse_feriado}
\usage{
parse_feriado(node)
}
\arguments{
\item{node}{Nó \code{xml2}.}
}
\value{
Um tibble de 1 linha.
}
\description{
Parseia uma estrutura \code{Feriado}.
}
