% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_response.R
\name{parse_generate_procedure}
\alias{parse_generate_procedure}
\title{Parse "gerarProcedimento" Response}
\usage{
parse_generate_procedure(resp_xml)
}
\arguments{
\item{resp_xml}{An \code{xml_document} object returned by \code{\link{sei_generate_procedure}}.}
}
\value{
A list with items \code{IdProcedimento}, \code{ProcedimentoFormatado}, and \code{LinkAcesso}.
  If any of these are missing, the list items will be \code{NA}.
}
\description{
Extracts key fields from the XML response of a \code{gerarProcedimento} call.
Assumes the SOAP response structure follows the SEI standard, with a
<return> node containing <IdProcedimento>, <ProcedimentoFormatado>, etc.
}
\examples{
\dontrun{
  resp_xml <- sei_generate_procedure(...your params...)
  info <- parse_generate_procedure(resp_xml)
  print(info$IdProcedimento)
  print(info$ProcedimentoFormatado)
  print(info$LinkAcesso)
}

}
