% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_listas.R
\name{parse_hipotese_legal}
\alias{parse_hipotese_legal}
\title{parse_hipotese_legal}
\usage{
parse_hipotese_legal(node)
}
\arguments{
\item{node}{Nó \code{xml2}.}
}
\value{
Um tibble de 1 linha.
}
\description{
Parseia uma estrutura \code{HipoteseLegal}.
}
