% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_consultas.R
\name{parse_imprensa_nacional}
\alias{parse_imprensa_nacional}
\title{parse_imprensa_nacional}
\usage{
parse_imprensa_nacional(node)
}
\arguments{
\item{node}{Nó \code{xml2}.}
}
\value{
Um tibble de 1 linha.
}
\description{
Parseia uma estrutura \code{PublicacaoImprensaNacional}.
}
