% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_response.R
\name{parse_interessado}
\alias{parse_interessado}
\title{parse_interessado}
\usage{
parse_interessado(nd_int)
}
\arguments{
\item{nd_int}{Um nó \code{xml2} de \code{<item>}.}
}
\value{
Um tibble de 1 linha com \code{Sigla} e \code{Nome}.
}
\description{
Parseia um \code{<item>} de \code{<Interessados>}.
}
