% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_listas.R
\name{parse_marcador}
\alias{parse_marcador}
\title{parse_marcador}
\usage{
parse_marcador(node)
}
\arguments{
\item{node}{Nó \code{xml2}.}
}
\value{
Um tibble de 1 linha.
}
\description{
Parseia uma estrutura \code{Marcador} (sem \code{Icone}, que é Base64).
}
