% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_response.R
\name{parse_observacao}
\alias{parse_observacao}
\title{parse_observacao}
\usage{
parse_observacao(nd_obs)
}
\arguments{
\item{nd_obs}{Um nó \code{xml2} de \code{<item>}.}
}
\value{
Um tibble de 1 linha com \code{Descricao}.
}
\description{
Parseia um \code{<item>} de \code{<Observacoes>}.
}
