% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sip.R
\name{parse_permissao}
\alias{parse_permissao}
\title{parse_permissao}
\usage{
parse_permissao(node)
}
\arguments{
\item{node}{Nó \code{xml2}.}
}
\value{
Um tibble de 1 linha.
}
\description{
Parseia uma estrutura \code{Permissao} do SIP.
}
