% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_response.R
\name{parse_procedimento_resumido}
\alias{parse_procedimento_resumido}
\title{parse_procedimento_resumido}
\usage{
parse_procedimento_resumido(nd_pr)
}
\arguments{
\item{nd_pr}{An \code{xml2} node for <item> in <ProcedimentosRelacionados> or <ProcedimentosAnexados>.}
}
\value{
A named list with \code{$IdProcedimento} and \code{$ProcedimentoFormatado}.
}
\description{
Parses one <ProcedimentoResumido> item, extracting \code{IdProcedimento} and
\code{ProcedimentoFormatado}.
}
\examples{
\dontrun{
  parse_procedimento_resumido(xml_item)
}

}
