% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_consultas.R
\name{parse_publicacao}
\alias{parse_publicacao}
\title{parse_publicacao}
\usage{
parse_publicacao(node)
}
\arguments{
\item{node}{Nó \code{xml2} da \code{Publicacao}.}
}
\value{
Um tibble de 1 linha.
}
\description{
Parseia uma estrutura \code{Publicacao}, com a \code{ImprensaNacional}
aninhada como coluna-lista.
}
