% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_escrita.R
\name{parse_retorno_geracao_procedimento}
\alias{parse_retorno_geracao_procedimento}
\title{parse_retorno_geracao_procedimento}
\usage{
parse_retorno_geracao_procedimento(doc)
}
\arguments{
\item{doc}{Um \code{xml2::xml_document}.}
}
\value{
Um tibble de 1 linha com \code{IdProcedimento}, \code{ProcedimentoFormatado},
\code{LinkAcesso} e \code{RetornoInclusaoDocumentos} (coluna-lista).
}
\description{
Parseia o retorno de \code{gerarProcedimento}
(\code{RetornoGeracaoProcedimento}).
}
