% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_escrita.R
\name{parse_retorno_inclusao_documento}
\alias{parse_retorno_inclusao_documento}
\title{parse_retorno_inclusao_documento}
\usage{
parse_retorno_inclusao_documento(doc)
}
\arguments{
\item{doc}{Um \code{xml2::xml_document}.}
}
\value{
Um tibble de 1 linha com \code{IdDocumento}, \code{DocumentoFormatado}, \code{LinkAcesso}.
}
\description{
Parseia o retorno de \code{incluirDocumento}
(\code{RetornoInclusaoDocumento}).
}
