% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_consultas.R
\name{parse_serie}
\alias{parse_serie}
\title{parse_serie}
\usage{
parse_serie(node)
}
\arguments{
\item{node}{Nó \code{xml2} da \code{Serie}.}
}
\value{
Um tibble de 1 linha com \code{IdSerie}, \code{Nome}, \code{Aplicabilidade}.
}
\description{
Parseia uma estrutura \code{Serie}.
}
