% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_listas.R
\name{parse_tipo_procedimento}
\alias{parse_tipo_procedimento}
\title{parse_tipo_procedimento}
\usage{
parse_tipo_procedimento(node)
}
\arguments{
\item{node}{Nó \code{xml2}.}
}
\value{
Um tibble de 1 linha.
}
\description{
Parseia uma estrutura \code{TipoProcedimento}.
}
