% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_consultas.R
\name{parse_unidade}
\alias{parse_unidade}
\title{parse_unidade}
\usage{
parse_unidade(node)
}
\arguments{
\item{node}{Nó \code{xml2} da \code{Unidade}.}
}
\value{
Um tibble de 1 linha com os campos da unidade.
}
\description{
Parseia uma estrutura \code{Unidade} (filho direto ou \verb{<item>}).
}
