% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_response.R
\name{parse_unidade_procedimento_aberto}
\alias{parse_unidade_procedimento_aberto}
\title{parse_unidade_procedimento_aberto}
\usage{
parse_unidade_procedimento_aberto(node_upa)
}
\arguments{
\item{node_upa}{An \code{xml2} node for one <item> in <UnidadesProcedimentoAberto>.}
}
\value{
A named list with \code{$Unidade} and \code{$UsuarioAtribuicao}, each a sub-list
  containing fields like \code{IdUnidade}, \code{Sigla}, \code{Descricao}.
}
\description{
Parses a node presumably within \code{<UnidadesProcedimentoAberto><item>...</item></UnidadesProcedimentoAberto>},
extracting \code{<Unidade>} and \code{<UsuarioAtribuicao>}.
}
\examples{
\dontrun{
  parse_unidade_procedimento_aberto(xml_item)
}

}
