% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rsei-package.R
\name{\%>\%}
\alias{\%>\%}
\title{Pipe operator}
\usage{
lhs \%>\% rhs
}
\arguments{
\item{lhs}{Um valor ou o objeto \code{magrittr} placeholder.}

\item{rhs}{Uma chamada de função usando a semântica do \code{magrittr}.}
}
\value{
O resultado de chamar \code{rhs(lhs)}.
}
\description{
Veja \code{magrittr::\link[magrittr:pipe]{\%>\%}} para detalhes.
}
\keyword{internal}
