% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/structures.R
\name{Procedimento}
\alias{Procedimento}
\title{Procedimento}
\usage{
Procedimento(
  IdTipoProcedimento,
  NumeroProtocolo = NULL,
  DataAutuacao = NULL,
  Especificacao = NULL,
  Assuntos = list(),
  Interessados = list(),
  Observacao = NULL,
  NivelAcesso = NULL,
  IdHipoteseLegal = NULL
)
}
\arguments{
\item{IdTipoProcedimento}{Character. ID of the process type.}

\item{NumeroProtocolo}{Character. Process number (optional).}

\item{DataAutuacao}{Character. Process creation date (optional).}

\item{Especificacao}{Character. Process specification (optional).}

\item{Assuntos}{A list of \code{\link{Assunto}} objects.}

\item{Interessados}{A list of \code{\link{Interessado}} objects.}

\item{Observacao}{Character. Unit observation (optional).}

\item{NivelAcesso}{"0"=public, "1"=restricted, "2"=secret, or NULL.}

\item{IdHipoteseLegal}{Character. If restricted/secret, link to a legal hypothesis.}
}
\value{
An S3 object of class "Procedimento".
}
\description{
Represents the "Procedimento" structure in SEI (process definition).
}
