% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/structures.R
\name{ProcedimentoResumido}
\alias{ProcedimentoResumido}
\title{ProcedimentoResumido}
\usage{
ProcedimentoResumido(
  IdTipoProcedimento,
  ProcedimentoFormatado,
  TipoProcedimento = NULL
)
}
\arguments{
\item{IdTipoProcedimento}{Character. ID of the process type.}

\item{ProcedimentoFormatado}{Character. Visible process number.}

\item{TipoProcedimento}{A \code{\link{TipoProcedimento}} object with details about the process type.}
}
\value{
An S3 object of class "ProcedimentoResumido".
}
\description{
Represents a "ProcedimentoResumido" structure (summary of a process).
}
