% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/structures.R
\name{ProtocoloBloco}
\alias{ProtocoloBloco}
\title{ProtocoloBloco}
\usage{
ProtocoloBloco(ProtocoloFormatado, Identificacao, Assinaturas = list())
}
\arguments{
\item{ProtocoloFormatado}{Character. Visible process or document number.}

\item{Identificacao}{Character. Type of the process or document.}

\item{Assinaturas}{A list of \code{\link{Assinatura}} objects (empty if none).}
}
\value{
An S3 object of class "ProtocoloBloco".
}
\description{
Represents a "ProtocoloBloco" structure (process/document in a block).
}
