% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/structures.R
\name{Publicacao}
\alias{Publicacao}
\title{Publicacao}
\usage{
Publicacao(
  IdPublicacao = NULL,
  IdDocumento = NULL,
  StaMotivo = NULL,
  Resumo = NULL,
  IdVeiculoPublicacao = NULL,
  NomeVeiculo = NULL,
  StaTipoVeiculo = NULL,
  Numero = NULL,
  DataDisponibilizacao = NULL,
  DataPublicacao = NULL,
  Estado = NULL,
  ImprensaNacional = NULL
)
}
\arguments{
\item{IdPublicacao}{Character. Internal ID of the publication.}

\item{IdDocumento}{Character. Internal ID of the associated document.}

\item{StaMotivo}{Character. "1"=Publication, "2"=Rectification, "3"=Republication, "4"=Apostilament.}

\item{Resumo}{Character. Summary text of the publication.}

\item{IdVeiculoPublicacao}{Character. Internal ID of the publication vehicle.}

\item{NomeVeiculo}{Character. Name of the publication vehicle.}

\item{StaTipoVeiculo}{Character. "I"=Internal, "E"=External, "M"=Module.}

\item{Numero}{Character. Publication number.}

\item{DataDisponibilizacao}{Character. Date of availability.}

\item{DataPublicacao}{Character. Date of publication.}

\item{Estado}{Character. "A"=Scheduled or "P"=Published.}

\item{ImprensaNacional}{A \code{\link{PublicacaoImprensaNacional}} object or NULL.}
}
\value{
An S3 object of class "Publicacao".
}
\description{
Represents the "Publicacao" structure in SEI (publication data).
}
