% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/structures.R
\name{PublicacaoImprensaNacional}
\alias{PublicacaoImprensaNacional}
\title{PublicacaoImprensaNacional}
\usage{
PublicacaoImprensaNacional(
  IdVeiculo = NULL,
  SiglaVeiculo = NULL,
  DescricaoVeiculo = NULL,
  Pagina = NULL,
  IdSecao = NULL,
  Secao = NULL,
  Data = NULL
)
}
\arguments{
\item{IdVeiculo}{Character. ID of the vehicle.}

\item{SiglaVeiculo}{Character. e.g., "DOU".}

\item{DescricaoVeiculo}{Character. e.g., "Diário Oficial da União".}

\item{Pagina}{Character. Page number of the publication.}

\item{IdSecao}{Character. ID of the section.}

\item{Secao}{Character. Section name.}

\item{Data}{Character. Publication date.}
}
\value{
An S3 object of class "PublicacaoImprensaNacional".
}
\description{
Represents the "PublicacaoImprensaNacional" structure in SEI.
}
