% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/escrita.R
\name{reabrir_processo}
\alias{reabrir_processo}
\title{reabrir_processo}
\usage{
reabrir_processo(
  protocolo_procedimento,
  config = sei_config(),
  verbose = FALSE
)
}
\arguments{
\item{protocolo_procedimento}{Character. Número do processo.}

\item{config}{Um objeto \code{\link[=sei_config]{sei_config()}}.}

\item{verbose}{Logical.}
}
\value{
\code{TRUE} em caso de sucesso.
}
\description{
Reabre um processo na unidade (\code{reabrirProcesso}).
}
