% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/escrita.R
\name{relacionar_processo}
\alias{relacionar_processo}
\title{relacionar_processo}
\usage{
relacionar_processo(
  protocolo1,
  protocolo2,
  config = sei_config(),
  verbose = FALSE
)
}
\arguments{
\item{protocolo1, protocolo2}{Números dos processos.}

\item{config}{Um objeto \code{\link[=sei_config]{sei_config()}}.}

\item{verbose}{Logical.}
}
\value{
\code{TRUE} em caso de sucesso.
}
\description{
Relaciona dois processos (\code{relacionarProcesso}). Bilateral.
}
