% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/structures.R
\name{Remetente}
\alias{Remetente}
\title{Remetente}
\usage{
Remetente(Sigla = NULL, Nome = NULL)
}
\arguments{
\item{Sigla}{Character. Participant sigla.}

\item{Nome}{Character. Participant name.}
}
\value{
An S3 object of class "Remetente".
}
\description{
Represents the "Remetente" structure in SEI (sender info).
}
