% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/escrita.R
\name{remover_controle_prazo}
\alias{remover_controle_prazo}
\title{remover_controle_prazo}
\usage{
remover_controle_prazo(
  protocolos_procedimentos,
  config = sei_config(),
  verbose = FALSE
)
}
\arguments{
\item{protocolos_procedimentos}{Vetor de números de processos.}

\item{config}{Um objeto \code{\link[=sei_config]{sei_config()}}.}

\item{verbose}{Logical.}
}
\value{
\code{TRUE} em caso de sucesso.
}
\description{
Remove o controle de prazo de processos (\code{removerControlePrazo}).
}
