% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/escrita.R
\name{remover_sobrestamento_processo}
\alias{remover_sobrestamento_processo}
\title{remover_sobrestamento_processo}
\usage{
remover_sobrestamento_processo(
  protocolo_procedimento,
  config = sei_config(),
  verbose = FALSE
)
}
\arguments{
\item{protocolo_procedimento}{Character. Número do processo.}

\item{config}{Um objeto \code{\link[=sei_config]{sei_config()}}.}

\item{verbose}{Logical.}
}
\value{
\code{TRUE} em caso de sucesso.
}
\description{
Remove o sobrestamento de um processo.
}
