% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sip.R
\name{replicar_permissao}
\alias{replicar_permissao}
\title{replicar_permissao}
\usage{
replicar_permissao(permissoes, config = sip_config(), verbose = FALSE)
}
\arguments{
\item{permissoes}{Lista de permissões; cada uma uma lista nomeada com campos
da estrutura \code{Permissao} (ao menos \code{StaOperacao}, \code{IdSistema}, \code{IdPerfil}).}

\item{config}{Um objeto \code{\link[=sip_config]{sip_config()}}.}

\item{verbose}{Logical.}
}
\value{
\code{TRUE} em caso de sucesso.
}
\description{
Replica (cadastra/altera/exclui) permissões no SIP (\code{replicarPermissao}).
}
