% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sip.R
\name{replicar_usuario}
\alias{replicar_usuario}
\title{replicar_usuario}
\usage{
replicar_usuario(
  usuarios,
  config = sip_config(),
  sin_considerar_orgao = "N",
  verbose = FALSE
)
}
\arguments{
\item{usuarios}{Lista de usuários; cada um uma lista nomeada (ao menos
\code{StaOperacao} e \code{IdOrigem}).}

\item{config}{Um objeto \code{\link[=sip_config]{sip_config()}}.}

\item{sin_considerar_orgao}{"S"/"N".}

\item{verbose}{Logical.}
}
\value{
\code{TRUE} em caso de sucesso.
}
\description{
Replica (cadastra/altera/desativa/reativa) usuários no SIP
(\code{replicarUsuario}).
}
