% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/escrita.R
\name{retirar_documento_bloco}
\alias{retirar_documento_bloco}
\title{retirar_documento_bloco}
\usage{
retirar_documento_bloco(
  id_bloco,
  protocolo_documento,
  config = sei_config(),
  verbose = FALSE
)
}
\arguments{
\item{id_bloco, protocolo_documento}{Identificadores.}

\item{config}{Um objeto \code{\link[=sei_config]{sei_config()}}.}

\item{verbose}{Logical.}
}
\value{
\code{TRUE} em caso de sucesso.
}
\description{
Retira um documento de um bloco (\code{retirarDocumentoBloco}).
}
