% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/escrita.R
\name{retirar_processo_bloco}
\alias{retirar_processo_bloco}
\title{retirar_processo_bloco}
\usage{
retirar_processo_bloco(
  id_bloco,
  protocolo_procedimento,
  config = sei_config(),
  verbose = FALSE
)
}
\arguments{
\item{id_bloco, protocolo_procedimento}{Identificadores.}

\item{config}{Um objeto \code{\link[=sei_config]{sei_config()}}.}

\item{verbose}{Logical.}
}
\value{
\code{TRUE} em caso de sucesso.
}
\description{
Retira um processo de um bloco (\code{retirarProcessoBloco}).
}
