% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{sei_listar_unidades}
\alias{sei_listar_unidades}
\title{List Units in SEI (listarUnidades)}
\usage{
sei_listar_unidades(
  sei_url,
  sigla_sistema,
  id_tipo_procedimento = NULL,
  id_serie = NULL,
  verbose = FALSE
)
}
\arguments{
\item{sei_url}{Character. The SEI Web Service endpoint,
e.g. "https://my-sei-server/sei/controlador_ws.php?servico=sei".}

\item{sigla_sistema}{Character. The system name/sigla, as registered in SEI.}

\item{id_tipo_procedimento}{Optional. The type of process configured
for the service.}

\item{id_serie}{Optional. The type of document configured
for the service.}

\item{verbose}{Logical. If TRUE, prints out the SOAP envelope
and the raw response for debugging purposes.}
}
\value{
An \code{xml_document} (from package \code{xml2}) with
  the raw SOAP response. This response typically contains multiple
  <Unidade> elements as described in the SEI documentation.
  For a more convenient R object, create a parser function to
  extract the relevant fields.
}
\description{
Calls the "listarUnidades" Web Service method, which returns a set of
unit structures according to the service's configured access. If
\code{IdTipoProcedimento} or \code{IdSerie} are supplied, they act as
additional filters.
}
\examples{
\dontrun{
  # Example usage
  resp_xml <- sei_listar_unidades(
    sei_url = "https://my-sei-server/sei/controlador_ws.php?servico=sei",
    sigla_sistema = "MySystem",
    identificacao_servico = "MySEIAccessKey",
    id_tipo_procedimento = "100000368",  # optional
    id_serie = "3",                     # optional
    verbose = TRUE
  )

  # If desired, parse resp_xml here to extract the <Unidade> info.
}

}
