% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/structures.R
\name{Serie}
\alias{Serie}
\title{Serie}
\usage{
Serie(IdSerie, Nome, Aplicabilidade = NULL)
}
\arguments{
\item{IdSerie}{Character. ID of the document type.}

\item{Nome}{Character. Name of the document type.}

\item{Aplicabilidade}{Character. "T"=internal/external, "I"=internal, "E"=external, "F"=forms.}
}
\value{
An S3 object of class "Serie".
}
\description{
Represents the "Serie" structure in SEI (document type).
}
