% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sip.R
\name{sip_config}
\alias{sip_config}
\title{Configuração de conexão SIP}
\usage{
sip_config(sip_url = NULL, chave_acesso = NULL, id_sistema = NULL)
}
\arguments{
\item{sip_url}{Character. Endpoint do SIP.}

\item{chave_acesso}{Character. Chave de Acesso do SIP.}

\item{id_sistema}{Character. Id do sistema no SIP.}
}
\value{
Um objeto de classe \code{sip_config}.
}
\description{
Objeto com os dados das chamadas SIP: URL do Web Service,
Chave de Acesso e \code{IdSistema}. Defaults resolvem de argumentos,
\verb{options(rsei.sip_*)}, variáveis \verb{RSEI_SIP_*} e o fallback embutido.
}
