% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/escrita.R
\name{sobrestar_processo}
\alias{sobrestar_processo}
\title{sobrestar_processo}
\usage{
sobrestar_processo(
  protocolo_procedimento,
  motivo,
  protocolo_vinculado = NULL,
  config = sei_config(),
  verbose = FALSE
)
}
\arguments{
\item{protocolo_procedimento}{Character. Número do processo.}

\item{motivo}{Character. Motivo do sobrestamento.}

\item{protocolo_vinculado}{Character opcional. Processo vinculado.}

\item{config}{Um objeto \code{\link[=sei_config]{sei_config()}}.}

\item{verbose}{Logical.}
}
\value{
\code{TRUE} em caso de sucesso.
}
\description{
Sobresta um processo (\code{sobrestarProcesso}).
}
