% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/credentials.R
\name{store_sei_credentials}
\alias{store_sei_credentials}
\title{Store SEI Credentials in Keyring}
\usage{
store_sei_credentials(service_name, username = NULL, password, ...)
}
\arguments{
\item{service_name}{Character string. The name of the service that identifies
this set of credentials (e.g., "SEI_WS").}

\item{username}{Optional. The username or system alias used for SEI, if applicable.
Defaults to "default_user" if not provided.}

\item{password}{The SEI access key or password to be stored.}

\item{...}{Other parameters you wish to store in the keyring (e.g., URL,
unit ID, etc.). These will be serialized as JSON under a separate username ("EXTRAS").}
}
\value{
Invisibly returns a logical value: `TRUE` if credentials were stored successfully,
  or `FALSE` otherwise.
}
\description{
Stores (or updates) credentials for the SEI web service in the `keyring`.
If the keyring package is not installed or if the store operation fails,
it simply displays a message instead of throwing an error.
}
\examples{
\dontrun{
store_sei_credentials(
  service_name = "SEI_WS",
  username = "MySystemSigla",
  password = "MySEIAccessKey",
  url = "https://myserver/sei/controlador_ws.php?servico=sei",
  unit_id = "100000969"
)
}

}
