% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/structures.R
\name{TipoConferencia}
\alias{TipoConferencia}
\title{TipoConferencia}
\usage{
TipoConferencia(IdTipoConferencia, Descricao)
}
\arguments{
\item{IdTipoConferencia}{Character. Identifier of the conference type.}

\item{Descricao}{Character. Description of the conference type.}
}
\value{
An S3 object of class "TipoConferencia".
}
\description{
Represents the "TipoConferencia" structure in SEI (conference type info).
}
