% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/structures.R
\name{TipoProcedimento}
\alias{TipoProcedimento}
\title{TipoProcedimento}
\usage{
TipoProcedimento(IdTipoProcedimento, Nome)
}
\arguments{
\item{IdTipoProcedimento}{Character. Identifier of the process type.}

\item{Nome}{Character. Name of the process type.}
}
\value{
An S3 object of class "TipoProcedimento".
}
\description{
Represents the "TipoProcedimento" structure in SEI (process type info).
}
