% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/structures.R
\name{Unidade}
\alias{Unidade}
\title{Unidade}
\usage{
Unidade(
  IdUnidade,
  Sigla,
  Descricao,
  SinProtocolo = NULL,
  SinArquivamento = NULL,
  SinOuvidoria = NULL
)
}
\arguments{
\item{IdUnidade}{Character. Identifier of the unit.}

\item{Sigla}{Character. Unit sigla.}

\item{Descricao}{Character. Unit description.}

\item{SinProtocolo}{"S" or "N" if it's a protocol unit.}

\item{SinArquivamento}{"S" or "N" if it's an archive unit.}

\item{SinOuvidoria}{"S" or "N" if it's an ombudsman unit.}
}
\value{
An S3 object of class "Unidade".
}
\description{
Represents the "Unidade" structure in SEI (organizational unit).
}
