% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/structures.R
\name{UnidadeProcedimentoAberto}
\alias{UnidadeProcedimentoAberto}
\title{UnidadeProcedimentoAberto}
\usage{
UnidadeProcedimentoAberto(Unidade, UsuarioAtribuicao = NULL)
}
\arguments{
\item{Unidade}{A \code{\link{Unidade}} object (the open unit).}

\item{UsuarioAtribuicao}{A \code{\link{Usuario}} object to which the process is assigned (optional).}
}
\value{
An S3 object of class "UnidadeProcedimentoAberto".
}
\description{
Represents the "UnidadeProcedimentoAberto" structure (where a process is open).
}
