% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/structures.R
\name{Usuario}
\alias{Usuario}
\title{Usuario}
\usage{
Usuario(IdUsuario, Sigla, Nome)
}
\arguments{
\item{IdUsuario}{Character. Identifier of the user.}

\item{Sigla}{Character. The user's sigla.}

\item{Nome}{Character. The user's full name.}
}
\value{
An S3 object of class "Usuario".
}
\description{
Represents the "Usuario" structure in SEI (user info).
}
