% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/soap_helpers.R
\name{build_parameters_xml}
\alias{build_parameters_xml}
\title{Build Simple XML for SOAP Parameters}
\usage{
build_parameters_xml(params_list)
}
\arguments{
\item{params_list}{A named list of parameters, e.g. \code{list(SiglaSistema = "X", IdUnidade = "123")} .}
}
\value{
A character string of XML that looks like:
  \code{<SiglaSistema>X</SiglaSistema><IdUnidade>123</IdUnidade>}.
}
\description{
Converts a named list (key-value pairs) to a simple XML string for inclusion
as parameters in the SOAP body.
}
\examples{
build_parameters_xml(list(SiglaSistema = "X", IdUnidade = "123"))

}
