% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/soap_helpers.R
\name{build_soap_envelope}
\alias{build_soap_envelope}
\title{Build a SOAP Envelope}
\usage{
build_soap_envelope(method, body_content)
}
\arguments{
\item{method}{Character string. The name of the SOAP method (e.g., "gerarProcedimento").}

\item{body_content}{Character string. The XML content to be placed inside the method tag.}
}
\value{
A character string containing the complete SOAP envelope XML.
}
\description{
Creates a basic SOAP envelope (SOAP 1.1) with the given `method` as the root
element in the `<soapenv:Body>`. The `body_content` should be a string containing
any child XML to be inserted.
}
\examples{
build_soap_envelope("myMethod", "<param>value</param>")

}
