% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consultaProcedimento.R
\name{consultar_procedimento}
\alias{consultar_procedimento}
\title{consultar_procedimento}
\usage{
consultar_procedimento(
  sei_url = "https://sei.pe.gov.br/sei/ws/SeiWS.php",
  sigla_sistema = "HORTENSIAS",
  identificacao_servico = "publicacao",
  id_unidade = "",
  protocolo_procedimento,
  sin_retornar_assuntos = "S",
  sin_retornar_interessados = "S",
  sin_retornar_observacoes = "S",
  sin_retornar_andamento_geracao = "S",
  sin_retornar_andamento_conclusao = "S",
  sin_retornar_ultimo_andamento = "S",
  sin_retornar_unidades_procedimento_aberto = "S",
  sin_retornar_procedimentos_relacionados = "S",
  sin_retornar_procedimentos_anexados = "S"
)
}
\arguments{
\item{sei_url}{Character. The SEI Web Service endpoint, typically something like
"https://sei.pe.gov.br/sei/ws/SeiWS.php".}

\item{sigla_sistema}{Character. The system sigla, e.g. "HORTENSIAS".}

\item{identificacao_servico}{Character. The key/service ID in SEI, e.g. "publicacao".}

\item{id_unidade}{Character. The SEI unit ID. If not needed, pass an empty string.}

\item{protocolo_procedimento}{Character. The visible process number, e.g. "0000000000.000003/2020-33".}

\item{sin_retornar_assuntos, sin_retornar_interessados, sin_retornar_observacoes, }{sin_retornar_andamento_geracao,sin_retornar_andamento_conclusao,sin_retornar_ultimo_andamento,
sin_retornar_unidades_procedimento_aberto,sin_retornar_procedimentos_relacionados,
sin_retornar_procedimentos_anexados
"S" or "N" to indicate whether you want those fields returned in the response.}
}
\value{
An xml2::xml_document (SOAP response) which you can parse further.
}
\description{
Sends a SOAP request to SEI (via \code{consultarProcedimento}) to retrieve
process data. Uses \code{glue::glue} to build the envelope, then performs
the request with \code{httr2}.
}
\examples{
\dontrun{
  doc <- consultar_procedimento(
    sei_url = "https://sei.pe.gov.br/sei/ws/SeiWS.php",
    sigla_sistema = "HORTENSIAS",
    identificacao_servico = "publicacao",
    id_unidade = "",
    protocolo_procedimento = "0000000000.000003/2020-33",
    sin_retornar_assuntos = "S",
    sin_retornar_interessados = "S",
    sin_retornar_observacoes = "S",
    sin_retornar_andamento_geracao = "S",
    sin_retornar_andamento_conclusao = "S",
    sin_retornar_ultimo_andamento = "S",
    sin_retornar_unidades_procedimento_aberto = "S",
    sin_retornar_procedimentos_relacionados = "S",
    sin_retornar_procedimentos_anexados = "S"
  )
  print(doc)
  # Then parse doc using your parser functions.
}

}
