% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/listarUnidades.R
\name{listar_unidades}
\alias{listar_unidades}
\title{listar_unidades}
\usage{
listar_unidades(
  sei_url = "https://sei.pe.gov.br/sei/ws/SeiWS.php",
  sigla_sistema = "HORTENSIAS",
  identificacao_servico = "publicacao"
)
}
\arguments{
\item{sei_url}{Character. The SEI Web Service endpoint, e.g. "https://sei.pe.gov.br/sei/ws/SeiWS.php".}

\item{sigla_sistema}{Character. The name/sigla of the system (e.g. "HORTENSIAS").}

\item{identificacao_servico}{Character. The service key in SEI (e.g. "publicacao").}
}
\value{
An \code{xml2::xml_document} containing the SOAP response.
  You can parse the relevant node (usually <parametros>) to get the array of units.
}
\description{
Sends a SOAP request to SEI (via \code{listarUnidades}) to retrieve
the list of available units (Unidades). This function replicates the
cURL command provided, using \code{glue} to build the envelope and
\code{httr2} to perform the request.
}
\examples{
\dontrun{
  doc <- listar_unidades(
    sei_url = "https://sei.pe.gov.br/sei/ws/SeiWS.php",
    sigla_sistema = "HORTENSIAS",
    identificacao_servico = "publicacao"
  )

  # Inspect raw XML:
  print(doc)

  # Then parse the result to extract <Unidade> elements or <parametros>.
  # For example:
  # nd_params <- xml_find_first(doc, "//*[local-name()='listarUnidadesResponse']/*[local-name()='parametros']")
  # unid_nodes <- xml_find_all(nd_params, ".//*[local-name()='Unidade']")
  # ... (extract with xml_text() or parse_unidade() function)
}

}
