% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_response.R
\name{parse_assunto}
\alias{parse_assunto}
\title{parse_assunto}
\usage{
parse_assunto(node_ass)
}
\arguments{
\item{node_ass}{Um nó \code{xml2} de \code{<item>}.}
}
\value{
Um tibble de 1 linha com \code{CodigoEstruturado} e \code{Descricao}.
}
\description{
Parseia um \code{<item>} de \code{<Assuntos>}.
}
