% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_response.R
\name{parse_consultar_procedimento_response}
\alias{parse_consultar_procedimento_response}
\title{parse_consultar_procedimento_response}
\usage{
parse_consultar_procedimento_response(doc)
}
\arguments{
\item{doc}{An \code{xml2::xml_document} containing the SOAP response
for the operation "consultarProcedimento".}
}
\value{
A named list with elements:
\itemize{
  \item \code{IdProcedimento}, \code{ProcedimentoFormatado}, \code{Especificacao}, \code{DataAutuacao},
        \code{LinkAcesso}, \code{NivelAcessoLocal}, \code{NivelAcessoGlobal}
  \item \code{TipoProcedimento} (list)
  \item \code{AndamentoGeracao}, \code{AndamentoConclusao}, \code{UltimoAndamento}
  \item \code{UnidadesProcedimentoAberto} (list of sub-lists)
  \item \code{Assuntos}, \code{Interessados}, \code{Observacoes},
        \code{ProcedimentosRelacionados}, \code{ProcedimentosAnexados}
}
}
\description{
The main parser that finds the <parametros> node under
<consultarProcedimentoResponse>, extracts top-level fields
(e.g., \code{IdProcedimento}, \code{Especificacao}, \code{NivelAcessoLocal}, etc.),
recodes \code{NivelAcessoLocal} and \code{NivelAcessoGlobal} using
\code{\link{recodify_access_level}}, and parses substructures like
\code{AndamentoGeracao}, \code{Assuntos}, \code{Interessados}, etc.
}
\examples{
\dontrun{
  resp_xml <- call_sei_api(..., method="consultarProcedimento")
  result <- parse_consultar_procedimento_response(resp_xml)
  str(result)
}

}
