% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_response.R
\name{parse_interessado}
\alias{parse_interessado}
\title{parse_interessado}
\usage{
parse_interessado(nd_int)
}
\arguments{
\item{nd_int}{An \code{xml2} node for <item> in <Interessados>.}
}
\value{
A named list with \code{$Sigla} and \code{$Nome}.
}
\description{
Parses one <Interessado> item, extracting \code{Sigla} and \code{Nome}.
}
\examples{
\dontrun{
  parse_interessado(xml_item)
}

}
