% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_response.R
\name{parse_observacao}
\alias{parse_observacao}
\title{parse_observacao}
\usage{
parse_observacao(nd_obs)
}
\arguments{
\item{nd_obs}{An \code{xml2} node for <item> in <Observacoes>.}
}
\value{
A named list with \code{$Descricao}.
}
\description{
Parses one <Observacao> item, extracting \code{Descricao} (and ignoring <Unidade> if present).
}
\examples{
\dontrun{
  parse_observacao(xml_item)
}

}
