% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_response.R
\name{parse_procedimento_resumido}
\alias{parse_procedimento_resumido}
\title{parse_procedimento_resumido}
\usage{
parse_procedimento_resumido(nd_pr)
}
\arguments{
\item{nd_pr}{Um nó \code{xml2} de \code{<item>}.}
}
\value{
Um tibble de 1 linha com \code{IdProcedimento} e \code{ProcedimentoFormatado}.
}
\description{
Parseia um \code{<item>} de \code{<ProcedimentosRelacionados>}
ou \code{<ProcedimentosAnexados>}.
}
