% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_response.R
\name{parse_unidade_procedimento_aberto}
\alias{parse_unidade_procedimento_aberto}
\title{parse_unidade_procedimento_aberto}
\usage{
parse_unidade_procedimento_aberto(node_upa)
}
\arguments{
\item{node_upa}{Um nó \code{xml2} de um \code{<item>}.}
}
\value{
Um tibble de 1 linha com os campos da unidade e do usuário de atribuição.
}
\description{
Parseia um \code{<item>} de \code{<UnidadesProcedimentoAberto>}, extraindo as
subestruturas \code{Unidade} e \code{UsuarioAtribuicao}.
}
