% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_response.R
\name{recodify_access_level}
\alias{recodify_access_level}
\title{Recodify SEI Access Level}
\usage{
recodify_access_level(val)
}
\arguments{
\item{val}{Character, geralmente "0", "1" ou "2".}
}
\value{
"público", "restrito", "sigiloso", ou o valor original.
}
\description{
Converte os códigos numéricos de nível de acesso do SEI em texto:
"0" -> "público", "1" -> "restrito", "2" -> "sigiloso". Valores ausentes,
vazios ou fora de \code{c("0","1","2")} são devolvidos inalterados.
}
\examples{
recodify_access_level("0")  # "público"
recodify_access_level("2")  # "sigiloso"

}
